review: guard deny on blocked status, drop double fetch, extract MCP action test factory
- deny is now a single bulk update guarded on the action still being
  blocked, so a concurrent approval is not clobbered
- clearActionRequiredForConversation variant avoids refetching the
  conversation already held by the caller
- extract AgentMCPActionFactory to replace copy-pasted test setup
