+++ front/lib/api/assistant/conversation/blocked_actions.test.ts
@@ -10,21 +10,16 @@ vi.mock("@app/lib/api/redis-hybrid-manager", () => ({
   }),
 }));
 
-import type { LightMCPToolConfigurationType } from "@app/lib/actions/mcp";
-import type { ToolExecutionStatus } from "@app/lib/actions/statuses";
 import { updateAgentMessageWithFinalStatus } from "@app/lib/api/assistant/conversation";
 import {
   clearActionRequiredIfNoBlockedActions,
   resolveBlockedActionsForTerminatedMessage,
 } from "@app/lib/api/assistant/conversation/blocked_actions";
 import type { Authenticator } from "@app/lib/auth";
-import { AgentStepContentToolExecutionModel } from "@app/lib/models/agent/actions/agent_step_content_tool_execution";
-import { AgentMCPActionModel } from "@app/lib/models/agent/actions/mcp";
-import { AgentStepContentModel } from "@app/lib/models/agent/agent_step_content";
 import { AgentMessageModel } from "@app/lib/models/agent/conversation";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
-import { generateRandomModelSId } from "@app/lib/resources/string_ids_server";
 import { AgentConfigurationFactory } from "@app/tests/utils/AgentConfigurationFactory";
+import { AgentMCPActionFactory } from "@app/tests/utils/AgentMCPActionFactory";
 import { ConversationFactory } from "@app/tests/utils/ConversationFactory";
 import { createResourceTest } from "@app/tests/utils/generic_resource_tests";
 import type { ConversationType } from "@app/types/assistant/conversation";
@@ -35,11 +30,8 @@ describe("blocked actions resolution", () => {
   let auth: Authenticator;
   let conversation: ConversationType;
 
-  let stepContentIndex = 0;
-
   beforeEach(async () => {
     vi.clearAllMocks();
-    stepContentIndex = 0;
 
     const setup = await createResourceTest({});
     workspace = setup.workspace;
@@ -57,89 +49,6 @@ describe("blocked actions resolution", () => {
       user: auth.getNonNullableUser().toJSON(),
     });
   });
-
-  /**
-   * Helper to create a blocked MCP action for testing.
-   */
-  async function createBlockedAction({
-    agentMessageId,
-    status = "blocked_validation_required",
-  }: {
-    agentMessageId: number;
-    status?: ToolExecutionStatus;
-  }) {
-    const functionCallId = generateRandomModelSId();
-    const currentIndex = stepContentIndex++;
-
-    const stepContent = await AgentStepContentModel.create({
-      workspaceId: workspace.id,
-      agentMessageId,
-      step: 1,
-      index: currentIndex,
-      version: 0,
-      type: "function_call",
-      value: {
-        type: "function_call",
-        value: {
-          id: functionCallId,
-          name: "test_tool",
-          arguments: "{}",
-        },
-      },
-    });
-
-    const toolConfiguration: LightMCPToolConfigurationType = {
-      id: 1,
-      sId: generateRandomModelSId(),
-      type: "mcp_configuration",
-      name: "test_tool",
-      dataSources: null,
-      tables: null,
-      childAgentId: null,
-      timeFrame: null,
-      jsonSchema: null,
-      additionalConfiguration: {},
-      mcpServerViewId: "test-server-view",
-      dustAppConfiguration: null,
-      secretName: null,
-      dustProject: null,
-      internalMCPServerId: null,
-      availability: "auto",
-      permission: "low",
-      toolServerId: "test-server",
-      retryPolicy: "no_retry",
-      originalName: "test_tool",
-      mcpServerName: "test_server",
-    };
-
-    const action = await AgentMCPActionModel.create({
-      workspaceId: workspace.id,
-      agentMessageId,
-      mcpServerConfigurationId: generateRandomModelSId(),
-      status,
-      citationsAllocated: 0,
-      augmentedInputs: {},
-      toolConfiguration,
-      stepContentId: stepContent.id,
-      stepContext: {
-        citationsCount: 0,
-        citationsOffset: 0,
-        resumeState: null,
-        retrievalTopK: 10,
-        websearchResultCount: 5,
-      },
-    });
-
-    await AgentStepContentToolExecutionModel.create({
-      workspaceId: workspace.id,
-      conversationId: conversation.id,
-      agentMessageId,
-      agentMCPActionId: action.id,
-      stepContentId: stepContent.id,
-    });
-
-    return { action, stepContent };
-  }
 
   async function createAgentMessageAtRank(rank: number) {
     const agentConfig = await AgentConfigurationFactory.createTestAgent(auth, {
@@ -179,7 +88,9 @@ describe("blocked actions resolution", () => {
     it("denies blocked actions and clears actionRequired when none remain", async () => {
       const { messageRow, agentMessageRowId } =
         await createAgentMessageAtRank(1);
-      const { action } = await createBlockedAction({
+      const { action } = await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
         agentMessageId: agentMessageRowId,
       });
 
@@ -209,13 +120,17 @@ describe("blocked actions resolution", () => {
     it("keeps actionRequired when another message still has a blocked action", async () => {
       const { messageRow, agentMessageRowId } =
         await createAgentMessageAtRank(1);
-      const { action } = await createBlockedAction({
+      const { action } = await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
         agentMessageId: agentMessageRowId,
       });
 
       const { agentMessageRowId: otherAgentMessageRowId } =
         await createAgentMessageAtRank(3);
-      const { action: otherAction } = await createBlockedAction({
+      const { action: otherAction } = await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
         agentMessageId: otherAgentMessageRowId,
       });
 
@@ -257,7 +172,11 @@ describe("blocked actions resolution", () => {
   describe("clearActionRequiredIfNoBlockedActions", () => {
     it("clears the flag when the only blocked action belongs to an unresumable message", async () => {
       const { agentMessageRowId } = await createAgentMessageAtRank(1);
-      await createBlockedAction({ agentMessageId: agentMessageRowId });
+      await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
+        agentMessageId: agentMessageRowId,
+      });
 
       // Simulate a legacy stuck conversation: the message was interrupted while its blocked
       // action was left pending.
@@ -278,7 +197,11 @@ describe("blocked actions resolution", () => {
 
     it("does not clear the flag when an actionable blocked action remains", async () => {
       const { agentMessageRowId } = await createAgentMessageAtRank(1);
-      await createBlockedAction({ agentMessageId: agentMessageRowId });
+      await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
+        agentMessageId: agentMessageRowId,
+      });
 
       await ConversationResource.markAsActionRequired(auth, { conversation });
 
@@ -300,7 +223,9 @@ describe("blocked actions resolution", () => {
         auth,
         { workspace, conversation, agentConfig }
       );
-      const { action } = await createBlockedAction({
+      const { action } = await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
         agentMessageId: agentMessage.agentMessageId,
       });
 
@@ -326,7 +251,9 @@ describe("blocked actions resolution", () => {
         auth,
         { workspace, conversation, agentConfig }
       );
-      const { action } = await createBlockedAction({
+      const { action } = await AgentMCPActionFactory.create({
+        workspace,
+        conversationId: conversation.id,
         agentMessageId: agentMessage.agentMessageId,
       });
